DLAI/Tutorial_Evaluating AI Agents/requirements.txt
# requirements file
# note which revision of python, for example 3.9.6
# in this file, insert all the pip install needs, include revision

#for example:
#torch==2.0.1
numpy==1.26.3
matplotlib

openinference-instrumentation==0.1.21
arize-phoenix-otel==0.7.1
arize-phoenix==7.0.0
openai==1.54.5
httpx<0.28
openinference-instrumentation-openai==0.1.18


sqlalchemy==2.0.36
aiosqlite==0.20.0
tqdm==4.65.0
duckdb==1.1.3

python-dotenv